code review - add some plot titles, increase number of PCs in PCA example
We'll want to show how much variance we're capturing with the number of PCs we choose to use. The main case we want to compare is -- how do we balance using all relevant predictor variables in a model with having a model that does not overfit? Answer: use as many PCs as you can without hitting any overfitting effects